# api/openapi.yaml
openapi: 3.0.3
info:
  title: Spacefleet API
  version: 0.1.0
  description: |
    HTTP API for the Spacefleet app. Regenerate consumers whenever this file changes.
servers:
  - url: /
paths:
  /api/health:
    get:
      summary: Liveness probe
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /api/ping:
    get:
      summary: Echo a greeting
      operationId: getPing
      parameters:
        - in: query
          name: name
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Greeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Greeting"
        default:
          $ref: "#/components/responses/Error"

  /api/cli/auth/approve:
    post:
      summary: Approve a pending CLI login (browser-initiated)
      description: |
        Called by the browser after the signed-in user confirms the CLI
        approval screen. Records a PKCE challenge and a human-friendly
        token name, returning a single-use exchange code that the browser
        redirects back to the CLI's localhost callback.
      operationId: approveCliAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CliApproveRequest"
      responses:
        "200":
          description: Approval recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliApproveResponse"
        default:
          $ref: "#/components/responses/Error"

  /api/cli/auth/exchange:
    post:
      summary: Exchange an approval code for a CLI token (public)
      description: |
        Called server-to-server by the CLI. Trades a one-time code plus the
        PKCE verifier for a long-lived bearer token. This endpoint is public
        because the CLI has no session yet when it calls it.
      operationId: exchangeCliAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CliExchangeRequest"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliExchangeResponse"
        default:
          $ref: "#/components/responses/Error"

  /api/cli/tokens:
    get:
      summary: List the caller's CLI tokens
      operationId: listCliTokens
      responses:
        "200":
          description: Token metadata (no plaintext)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliTokenList"
        default:
          $ref: "#/components/responses/Error"

  /api/cli/tokens/{id}:
    delete:
      summary: Revoke a CLI token
      operationId: revokeCliToken
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Revoked (idempotent)
        default:
          $ref: "#/components/responses/Error"

  /api/orgs/{slug}/github/installations/start:
    post:
      summary: Begin a GitHub App install handshake
      description: |
        Issues a one-time CSRF state and returns the URL the browser should
        send the user to. After install, GitHub redirects back to the App's
        configured setup URL with `installation_id` and `state`; the SPA
        finishes the handshake via `/api/github/installations/complete`.
      operationId: startGithubInstall
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      responses:
        "200":
          description: Install URL ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubInstallStart"
        default:
          $ref: "#/components/responses/Error"

  /api/github/installations/complete:
    post:
      summary: Finalize a GitHub App install handshake (Clerk session only)
      description: |
        Called by the SPA setup-callback page with the `state` and
        `installation_id` returned from GitHub. Verifies the state, fetches
        the installation from GitHub via App-level auth, and persists it
        scoped to the org that started the handshake. Not org-prefixed
        because the org is recovered from the state itself.
      operationId: completeGithubInstall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubInstallCompleteRequest"
      responses:
        "200":
          description: Installation persisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubInstallCompleteResponse"
        default:
          $ref: "#/components/responses/Error"

  /api/orgs/{slug}/github/installations:
    get:
      summary: List GitHub App installations connected to this org
      operationId: listGithubInstallations
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      responses:
        "200":
          description: Installations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubInstallationList"
        default:
          $ref: "#/components/responses/Error"

  /api/orgs/{slug}/github/installations/{installationId}:
    delete:
      summary: Disconnect a GitHub App installation from this org
      description: |
        Drops Spacefleet's record of the installation. Does not uninstall
        the GitHub App on github.com — that's the user's choice.
      operationId: deleteGithubInstallation
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/InstallationID"
      responses:
        "204":
          description: Disconnected (idempotent)
        default:
          $ref: "#/components/responses/Error"

  /api/orgs/{slug}/github/installations/{installationId}/repositories:
    get:
      summary: List repositories accessible to a GitHub App installation
      description: |
        Mints a fresh installation access token and enumerates the repos
        the App has been granted on. This is what the build pipeline will
        clone from.
      operationId: listGithubInstallationRepositories
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/InstallationID"
      responses:
        "200":
          description: Repositories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubRepositoryList"
        default:
          $ref: "#/components/responses/Error"

  /api/cli/whoami:
    get:
      summary: Return the user associated with the CLI token
      description: |
        Smoke-test endpoint for the CLI — confirms a token is valid and
        reports whose identity it carries. Clerk-authenticated calls also
        work; they simply echo the Clerk session's user id.
      operationId: cliWhoami
      responses:
        "200":
          description: Identity echo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliWhoami"
        default:
          $ref: "#/components/responses/Error"

components:
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    Greeting:
      type: object
      required: [message]
      properties:
        message:
          type: string
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string

    CliApproveRequest:
      type: object
      required: [name, challenge]
      properties:
        name:
          type: string
          description: User-chosen label for the token (e.g. hostname).
        challenge:
          type: string
          description: Base64url(sha256(verifier)) — the CLI's PKCE challenge.

    CliApproveResponse:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: Single-use exchange code, valid for ~5 minutes.

    CliExchangeRequest:
      type: object
      required: [code, verifier]
      properties:
        code:
          type: string
        verifier:
          type: string

    CliExchangeResponse:
      type: object
      required: [token, name, expires_at]
      properties:
        token:
          type: string
          description: The bearer token. Returned exactly once — not recoverable later.
        name:
          type: string
        expires_at:
          type: string
          format: date-time

    CliToken:
      type: object
      required: [id, name, created_at, expires_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true

    CliTokenList:
      type: object
      required: [tokens]
      properties:
        tokens:
          type: array
          items:
            $ref: "#/components/schemas/CliToken"

    CliWhoami:
      type: object
      required: [user_id, source]
      properties:
        user_id:
          type: string
        source:
          type: string
          enum: [clerk, cli]

    GithubInstallStart:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: |
            Send the user here to install the App. State is embedded in the
            URL; the caller does not need to track it.

    GithubInstallCompleteRequest:
      type: object
      required: [state, installation_id]
      properties:
        state:
          type: string
        installation_id:
          type: integer
          format: int64

    GithubInstallCompleteResponse:
      type: object
      required: [installation, org_slug]
      properties:
        installation:
          $ref: "#/components/schemas/GithubInstallation"
        org_slug:
          type: string
          description: The org the installation was bound to. Use it to redirect.

    GithubInstallation:
      type: object
      required:
        [
          id,
          installation_id,
          account_login,
          account_type,
          account_id,
          created_at,
          updated_at,
        ]
      properties:
        id:
          type: string
          format: uuid
        installation_id:
          type: integer
          format: int64
        account_login:
          type: string
        account_type:
          type: string
          description: '"User" or "Organization" (mirrors GitHub).'
        account_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        suspended_at:
          type: string
          format: date-time
          nullable: true

    GithubInstallationList:
      type: object
      required: [installations]
      properties:
        installations:
          type: array
          items:
            $ref: "#/components/schemas/GithubInstallation"

    GithubRepository:
      type: object
      required: [id, name, full_name, private, default_branch, html_url]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        full_name:
          type: string
        private:
          type: boolean
        default_branch:
          type: string
        html_url:
          type: string
          format: uri

    GithubRepositoryList:
      type: object
      required: [repositories]
      properties:
        repositories:
          type: array
          items:
            $ref: "#/components/schemas/GithubRepository"

  parameters:
    OrgSlug:
      in: path
      name: slug
      required: true
      schema:
        type: string
    InstallationID:
      in: path
      name: installationId
      required: true
      schema:
        type: integer
        format: int64

  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"