# ent/githubinstallation.go
// Code generated by ent, DO NOT EDIT.

package ent

import (
	"fmt"
	"strings"
	"time"

	"entgo.io/ent"
	"entgo.io/ent/dialect/sql"
	"github.com/google/uuid"
	"github.com/spacefleet/app/ent/githubinstallation"
)

// GithubInstallation is the model entity for the GithubInstallation schema.
type GithubInstallation struct {
	config `json:"-"`
	// ID of the ent.
	ID uuid.UUID `json:"id,omitempty"`
	// OrgSlug holds the value of the "org_slug" field.
	OrgSlug string `json:"org_slug,omitempty"`
	// InstallationID holds the value of the "installation_id" field.
	InstallationID int64 `json:"installation_id,omitempty"`
	// AccountLogin holds the value of the "account_login" field.
	AccountLogin string `json:"account_login,omitempty"`
	// AccountType holds the value of the "account_type" field.
	AccountType string `json:"account_type,omitempty"`
	// AccountID holds the value of the "account_id" field.
	AccountID int64 `json:"account_id,omitempty"`
	// CreatedAt holds the value of the "created_at" field.
	CreatedAt time.Time `json:"created_at,omitempty"`
	// UpdatedAt holds the value of the "updated_at" field.
	UpdatedAt time.Time `json:"updated_at,omitempty"`
	// SuspendedAt holds the value of the "suspended_at" field.
	SuspendedAt  *time.Time `json:"suspended_at,omitempty"`
	selectValues sql.SelectValues
}

// scanValues returns the types for scanning values from sql.Rows.
func (*GithubInstallation) scanValues(columns []string) ([]any, error) {
	values := make([]any, len(columns))
	for i := range columns {
		switch columns[i] {
		case githubinstallation.FieldInstallationID, githubinstallation.FieldAccountID:
			values[i] = new(sql.NullInt64)
		case githubinstallation.FieldOrgSlug, githubinstallation.FieldAccountLogin, githubinstallation.FieldAccountType:
			values[i] = new(sql.NullString)
		case githubinstallation.FieldCreatedAt, githubinstallation.FieldUpdatedAt, githubinstallation.FieldSuspendedAt:
			values[i] = new(sql.NullTime)
		case githubinstallation.FieldID:
			values[i] = new(uuid.UUID)
		default:
			values[i] = new(sql.UnknownType)
		}
	}
	return values, nil
}

// assignValues assigns the values that were returned from sql.Rows (after scanning)
// to the GithubInstallation fields.
func (_m *GithubInstallation) assignValues(columns []string, values []any) error {
	if m, n := len(values), len(columns); m < n {
		return fmt.Errorf("mismatch number of scan values: %d != %d", m, n)
	}
	for i := range columns {
		switch columns[i] {
		case githubinstallation.FieldID:
			if value, ok := values[i].(*uuid.UUID); !ok {
				return fmt.Errorf("unexpected type %T for field id", values[i])
			} else if value != nil {
				_m.ID = *value
			}
		case githubinstallation.FieldOrgSlug:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field org_slug", values[i])
			} else if value.Valid {
				_m.OrgSlug = value.String
			}
		case githubinstallation.FieldInstallationID:
			if value, ok := values[i].(*sql.NullInt64); !ok {
				return fmt.Errorf("unexpected type %T for field installation_id", values[i])
			} else if value.Valid {
				_m.InstallationID = value.Int64
			}
		case githubinstallation.FieldAccountLogin:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field account_login", values[i])
			} else if value.Valid {
				_m.AccountLogin = value.String
			}
		case githubinstallation.FieldAccountType:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field account_type", values[i])
			} else if value.Valid {
				_m.AccountType = value.String
			}
		case githubinstallation.FieldAccountID:
			if value, ok := values[i].(*sql.NullInt64); !ok {
				return fmt.Errorf("unexpected type %T for field account_id", values[i])
			} else if value.Valid {
				_m.AccountID = value.Int64
			}
		case githubinstallation.FieldCreatedAt:
			if value, ok := values[i].(*sql.NullTime); !ok {
				return fmt.Errorf("unexpected type %T for field created_at", values[i])
			} else if value.Valid {
				_m.CreatedAt = value.Time
			}
		case githubinstallation.FieldUpdatedAt:
			if value, ok := values[i].(*sql.NullTime); !ok {
				return fmt.Errorf("unexpected type %T for field updated_at", values[i])
			} else if value.Valid {
				_m.UpdatedAt = value.Time
			}
		case githubinstallation.FieldSuspendedAt:
			if value, ok := values[i].(*sql.NullTime); !ok {
				return fmt.Errorf("unexpected type %T for field suspended_at", values[i])
			} else if value.Valid {
				_m.SuspendedAt = new(time.Time)
				*_m.SuspendedAt = value.Time
			}
		default:
			_m.selectValues.Set(columns[i], values[i])
		}
	}
	return nil
}

// Value returns the ent.Value that was dynamically selected and assigned to the GithubInstallation.
// This includes values selected through modifiers, order, etc.
func (_m *GithubInstallation) Value(name string) (ent.Value, error) {
	return _m.selectValues.Get(name)
}

// Update returns a builder for updating this GithubInstallation.
// Note that you need to call GithubInstallation.Unwrap() before calling this method if this GithubInstallation
// was returned from a transaction, and the transaction was committed or rolled back.
func (_m *GithubInstallation) Update() *GithubInstallationUpdateOne {
	return NewGithubInstallationClient(_m.config).UpdateOne(_m)
}

// Unwrap unwraps the GithubInstallation entity that was returned from a transaction after it was closed,
// so that all future queries will be executed through the driver which created the transaction.
func (_m *GithubInstallation) Unwrap() *GithubInstallation {
	_tx, ok := _m.config.driver.(*txDriver)
	if !ok {
		panic("ent: GithubInstallation is not a transactional entity")
	}
	_m.config.driver = _tx.drv
	return _m
}

// String implements the fmt.Stringer.
func (_m *GithubInstallation) String() string {
	var builder strings.Builder
	builder.WriteString("GithubInstallation(")
	builder.WriteString(fmt.Sprintf("id=%v, ", _m.ID))
	builder.WriteString("org_slug=")
	builder.WriteString(_m.OrgSlug)
	builder.WriteString(", ")
	builder.WriteString("installation_id=")
	builder.WriteString(fmt.Sprintf("%v", _m.InstallationID))
	builder.WriteString(", ")
	builder.WriteString("account_login=")
	builder.WriteString(_m.AccountLogin)
	builder.WriteString(", ")
	builder.WriteString("account_type=")
	builder.WriteString(_m.AccountType)
	builder.WriteString(", ")
	builder.WriteString("account_id=")
	builder.WriteString(fmt.Sprintf("%v", _m.AccountID))
	builder.WriteString(", ")
	builder.WriteString("created_at=")
	builder.WriteString(_m.CreatedAt.Format(time.ANSIC))
	builder.WriteString(", ")
	builder.WriteString("updated_at=")
	builder.WriteString(_m.UpdatedAt.Format(time.ANSIC))
	builder.WriteString(", ")
	if v := _m.SuspendedAt; v != nil {
		builder.WriteString("suspended_at=")
		builder.WriteString(v.Format(time.ANSIC))
	}
	builder.WriteByte(')')
	return builder.String()
}

// GithubInstallations is a parsable slice of GithubInstallation.
type GithubInstallations []*GithubInstallation